# Optimize internal multilabel score